+++ src/kanbanwave/KanbanStorageProvider.tsx
@@ -5,14 +5,13 @@ import {
   useMemo,
   useSyncExternalStore
 } from 'react';
+import { type KanbanExternalStore, makeKanbanExternalStore } from './storage';
 import {
   KanbanBoardStorage,
   KanbanCardStorage,
-  KanbanExternalStore,
   KanbanListStorage,
-  KanbanStorage,
-  makeKanbanExternalStore
-} from 'kanbanwave';
+  KanbanStorage
+} from './types';
 
 export type KanbanStorageContextValue = {
   board: KanbanExternalStore<KanbanBoardStorage>;
