+++ src/kanbanwave/index.ts
@@ -1,2 +1,11 @@
+
 export * from './storage';
 export * from './types';
+
+export {
+  default as KanbanStorageProvider,
+  useKanbanBoard,
+  useKanbanList,
+  useKanbanCard,
+  type KanbanStorageContextValue
+} from './KanbanStorageProvider';
